a fix for jurre-brishti forecasts generator

+++ fastcgan/jobs/proxy_sync.py
@@ -1,5 +1,5 @@
 from argparse import ArgumentParser
-from typing import Any, Literal
+from typing import Literal
 
 import requests
 from bs4 import BeautifulSoup
@@ -14,8 +14,9 @@ def crawl_http_dataset_links(data_page: str) -> list[str]:
     r = requests.get(data_page, allow_redirects=True)
     if r.status_code == 200:
         soup = BeautifulSoup(r.text, features="html.parser")
-        data_files.extend([f"{data_page}{a['href']}" for a in soup.find_all("a") if "../" not in a])
-        logger.info(f"crawled a total of {len(data_files)} data files from {data_page}")
+        data_files.extend(
+            [f"{data_page}{a['href']}" for a in soup.find_all("a") if "../" not in a]
+        )
     else:
         logger.warning(
             f"failed to crawl links from {data_page} with status code {r.status_code} and response text {r.text}"
@@ -23,10 +24,34 @@ def crawl_http_dataset_links(data_page: str) -> list[str]:
     return data_files
 
 
-def make_dataset_path(dataset_url: str, data_source: str, trim_part: str | None = "") -> None:
+def deep_crawl_http_dataset_links(
+    data_page: str, data_ext: str | None = "nc", links: set[str] | None = set()
+) -> list[str]:
+    r = requests.get(data_page, allow_redirects=True)
+    if r.status_code == 200:
+        soup = BeautifulSoup(r.text, features="html.parser")
+        for a in soup.find_all("a"):
+            href = f"{data_page}/{a['href']}"
+            if href.endswith(data_ext):
+                links.add(href)
+            elif "../" not in href and "//ICPAC" not in href and href.endswith("/"):
+                links = deep_crawl_http_dataset_links(data_page=href, links=links)
+        logger.info(f"crawled a total of {len(links)} data files from {data_page}")
+    else:
+        logger.warning(
+            f"failed to crawl links from {data_page} with status code {r.status_code} and response text {r.text}"
+        )
+    return links
+
+
+def make_dataset_path(
+    dataset_url: str, data_source: str, trim_part: str | None = ""
+) -> None:
     dir_tree = "/".join(dataset_url.replace(trim_part, "").split("/"))
     forecasts_dir = get_data_store_path(source=data_source)
-    data_dir = forecasts_dir / f"{dir_tree.replace('%20', ' ') if dir_tree != '/' else ''}"
+    data_dir = (
+        forecasts_dir / f"{dir_tree.replace('%20', ' ') if dir_tree != '/' else ''}"
+    )
     if not data_dir.exists():
         data_dir.mkdir(parents=True)
 
@@ -53,76 +78,39 @@ def retrieve_cgan_data_links(
     return datasets_links
 
 
-def download_ens_dataset(source: str, year: str, month: str, day: str, start_time: str, valid_time: str):
+def download_ens_dataset(source: str, link: str):
+    file_name = link.split("/")[-1]
+    data_date = file_name.split("_")[1]
     data_dir = get_data_store_path(source=f"{source}-count")
-    model_path = "Jurre_brishti_counts" if source == "jurre-brishti" else "Mvua_kubwa_counts"
-    dwnld_link = (
-        f"http://megacorr.dynu.net/ICPAC/cGAN_examplePlots/data/{model_path}/{year}"
-        + f"/counts_{year}{month}{day}_{start_time}_{valid_time}h.nc"
-    )
-    destination = data_dir / year / month
+    destination = data_dir / data_date[:4] / data_date[4:6]
     if not destination.exists():
         destination.mkdir(parents=True)
-    logger.debug(f"trying download of {dwnld_link}")
-    try:
-        file_path = destination / f"counts_{year}{month}{day}_{start_time}_{valid_time}h.nc"
-        with requests.get(dwnld_link, stream=True) as r:
-            logger.debug(f"downloading {dwnld_link} into {destination}")
+    file_path = destination / file_name
+    if not file_path.exists():
+        logger.debug(f"trying download of {link}")
+        try:
+            with requests.get(link, stream=True) as r:
+                logger.debug(f"downloading {link} into {file_path}")
 
-            if r.status_code == 200:
-                with file_path.open(mode="wb") as f:
-                    f.write(r.content)
-                logger.info(f"Finished downloading {dwnld_link}.\t" + f"Data stream was saved in {file_path.name}")
-            else:
-                logger.error(f"failed to download dataset file {dwnld_link} " + f"with http response {r.text}")
-
-    except Exception as err:
-        logger.error(f"failed to download {dwnld_link} with error {err}")
-
-
-def retrieve_ens_counts_datasets_for_date(
-    source: Literal["jurre-brishti", "mvua-kubwa"], data_date: str, forecast_dates: list[dict[str, Any]] | None = None
-):
-    logger.debug(f"received cgan ensemble counts retrieval task for {source} model {data_date}")
-    year, month, day = (str(int(value)) for value in data_date.split("-"))
-    if forecast_dates is None:
-        model_path = "Jurre_brishti_counts" if source == "jurre-brishti" else "Mvua_kubwa_counts"
-        dates_json = f"http://megacorr.dynu.net/ICPAC/cGAN_examplePlots/data/{model_path}/available_dates.json"
-        r = requests.get(dates_json)
-        if r.status_code == 200:
-            forecast_dates = r.json()
-    if forecast_dates is not None:
-        for start_time in reversed(forecast_dates[year][month][day].keys()):
-            for valid_time in reversed(forecast_dates[year][month][day][start_time]):
-                download_ens_dataset(
-                    source=source,
-                    year=year,
-                    month=month.rjust(2, "0"),
-                    day=day.rjust(2, "0"),
-                    start_time=str(start_time).rjust(2, "0"),
-                    valid_time=valid_time,
-                )
-
-
-def retrieve_ens_counts_datasets(source: Literal["jurre-brishti", "mvua-kubwa"]):
-    model_path = "Jurre_brishti_counts" if source == "jurre-brishti" else "Mvua_kubwa_counts"
-    dates_json = f"http://megacorr.dynu.net/ICPAC/cGAN_examplePlots/data/{model_path}/available_dates.json"
-    r = requests.get(dates_json)
-    if r.status_code == 200:
-        forecast_dates = r.json()
-        for year in reversed(forecast_dates.keys()):
-            for month in reversed(forecast_dates[year].keys()):
-                for day in reversed(forecast_dates[year][month].keys()):
-                    retrieve_ens_counts_datasets_for_date(
-                        source=source,
-                        forecast_dates=forecast_dates,
-                        data_date=f"{year}-{str(month).rjust(2, '0')}-{str(day).rjust(2, '0')}",
+                if r.status_code == 200:
+                    with file_path.open(mode="wb") as f:
+                        f.write(r.content)
+                    logger.info(
+                        f"Finished downloading {link}.\t"
+                        + f"Data stream was saved into {file_path.name}"
                     )
+                else:
+                    logger.error(
+                        f"failed to download dataset file {link} "
+                        + f"with http response {r.text}"
+                    )
+
+        except Exception as err:
+            logger.error(f"failed to download {link} with error {err}")
 
 
 def sync_data_source(
     sources: str,
-    data_date: str | None = None,
     start_month: int | None = 1,
     final_month: int | None = 12,
     year: int | None = 2024,
@@ -130,10 +118,20 @@ def sync_data_source(
 ) -> None:
     for source in sources.split(","):
         if source == "mvua-kubwa" or source == "jurre-brishti":
-            if data_date is not None:
-                retrieve_ens_counts_datasets_for_date(data_date=data_date, source=source)
-            else:
-                retrieve_ens_counts_datasets(source)
+            provider_url = "http://megacorr.dynu.net/ICPAC/cGAN_examplePlots/data"
+            data_path = (
+                "Jurre_brishti_counts"
+                if source == "jurre-brishti"
+                else "Mvua_kubwa_counts"
+            )
+            links = deep_crawl_http_dataset_links(
+                data_page=f"{provider_url}/{data_path}"
+            )
+            logger.info(
+                f"crawled a total of {len(links)} data files from {provider_url}"
+            )
+            for link in links:
+                download_ens_dataset(source=source, link=link)
         else:
             data_urls = retrieve_cgan_data_links(
                 start_month=start_month,
@@ -145,8 +143,12 @@ def sync_data_source(
             for datafile_url in data_urls:
                 logger.debug(f"trying download of {datafile_url}")
                 try:
-                    relative_path = datafile_url.replace(provider_url, "").replace(f"/{source}", "")
-                    destination = get_data_store_path(source=source) / f"{relative_path}".replace("%20", " ")
+                    relative_path = datafile_url.replace(provider_url, "").replace(
+                        f"/{source}", ""
+                    )
+                    destination = get_data_store_path(
+                        source=source
+                    ) / f"{relative_path}".replace("%20", " ")
                     with requests.get(datafile_url, stream=True) as r:
                         logger.debug(f"downloading {datafile_url} into {destination}")
 
@@ -154,10 +156,14 @@ def sync_data_source(
                             with destination.open(mode="wb") as f:
                                 f.write(r.content)
                         else:
-                            logger.error(f"failed to download dataset file {datafile_url} with error {r.text}")
+                            logger.error(
+                                f"failed to download dataset file {datafile_url} with error {r.text}"
+                            )
                             return None
 
-                        logger.debug(f"Finished downloading {datafile_url}.\nData stream was saved in {f.name}")
+                        logger.debug(
+                            f"Finished downloading {datafile_url}.\nData stream was saved in {f.name}"
+                        )
 
                 except Exception as e:
                     logger.error(f"Error *{e}*")
@@ -170,7 +176,9 @@ if __name__ == "__main__":
         description="a program for downloading data from online cGAN sources",
         usage="python download.py -y <year> -sm <start-month> -fm <final-month>",
     )
-    parser.add_argument("-y", "--year", dest="year", type=int, default=2024, help="data year")
+    parser.add_argument(
+        "-y", "--year", dest="year", type=int, default=2024, help="data year"
+    )
     parser.add_argument(
         "-sm",
         "--start-month",
